requirements.txt
flask==2.0.1
google-generativeai==0.3.2 